+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/artwork/ArtworkService.java
@@ -1,8 +1,8 @@
 package com.startingblue.fourtooncookie.artwork;
 
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
-import com.startingblue.fourtooncookie.artwork.dto.request.ArtworkSaveRequest;
-import com.startingblue.fourtooncookie.artwork.dto.request.ArtworkUpdateRequest;
+import com.startingblue.fourtooncookie.artwork.dto.ArtworkSaveRequest;
+import com.startingblue.fourtooncookie.artwork.dto.ArtworkUpdateRequest;
 import com.startingblue.fourtooncookie.artwork.exception.ArtworkDuplicateException;
 import com.startingblue.fourtooncookie.artwork.exception.ArtworkNotFoundException;
 import lombok.RequiredArgsConstructor;
